Insurance-Claim-Help-Microservice-docker-vps-deploy/requirements.txt
absl-py==2.4.0
addict==2.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
av==17.0.0
babel==2.18.0
beautifulsoup4==4.14.3
bidict==0.23.1
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
comet_ml==3.57.3
comm==0.2.3
ConfigArgParse==1.7.5
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dash==4.0.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
dill==0.4.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
docstring_parser==0.17.0
dulwich==0.25.2
everett==3.1.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
fire==0.7.1
Flask==3.1.3
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.2.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.3.9
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.2
lark==1.3.1
lazy-loader==0.5
lightning-utilities==0.15.3
lxml==6.0.2
mapbox_earcut==2.0.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multiprocess==0.70.19
mysql-connector-python==9.6.0
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nodeenv==1.10.0
notebook_shim==0.2.4
numpy>=2.0.0,<3.0.0
opencv-python>=4.13.0.92
opencv-python-headless==4.10.0.84
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.6
pathos==0.3.5
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
plyfile==1.1.3
polars==1.38.1
polars-runtime-32==1.38.1
pox==0.3.7
ppft==1.7.8
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf>=4.25.8,<7.0.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybind11==3.0.2
pycocotools==2.0.11
pycollada==0.9.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyliblzfse==0.4.1
pyngrok==7.5.1
pyparsing==3.3.2
pyquaternion==0.9.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-engineio==4.13.1
python-json-logger==4.0.0
python-socketio==5.16.1
pytorch-msssim==1.0.0
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
reportlab==4.4.10
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
Send2Trash==2.1.0
sentry-sdk==2.55.0
setuptools==82.0.0
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
splines==0.3.0
sqlparse==0.5.5
whitenoise==6.9.0
stack-data==0.6.3
svg.path==7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorly==0.9.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==0.6.7
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
torch_fidelity==0.4.0
torchmetrics==1.9.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
trimesh==4.11.4
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.10
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.4.21
ultralytics-thop==2.0.18
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
vhacdx==0.0.10
vine==5.1.0
viser==0.2.7
wadler_lindig==0.1.7
wandb==0.25.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
wrapt==2.1.2
wsproto==1.3.2
wurlitzer==3.1.1
xatlas==0.0.11
xxhash==3.6.0
zipp==3.23.0
